move callback handled to Waypoint
