+++ src/index.jsx
@@ -1,5 +1,5 @@
 import Waypoint from './waypoint';
 export default Waypoint;
 
-import waypoint from './waypoint-hoc';
-export { waypoint };
+import waypoint, { POSITIONS } from './waypoint-hoc';
+export { waypoint, POSITIONS };
